backend/requirements/base.txt
Django==4.2.1
django-cors-headers==3.13.0
djangorestframework==3.14.0
django-environ==0.10.0
django-filter==23.2
djangorestframework-simplejwt==5.2.2
drf-flex-fields==1.0.2
huey==2.4.5
drf-spectacular==0.27.0
Pillow==10.1.0
drf-flex-fields==1.0.2
django-safedelete==1.4.0
requests==2.32.3